shoppingList/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-crispy-forms==2.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Markdown==3.6
mysqlclient==2.2.4
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
